Write and flush top/bottom-screen descriptors
A descriptor is consumed immediately when bound to a command buffer and
cannot be updated after unless we mark the descriptors with
`UPDATE_AFTER_BIND` flags and such.

For now, just flush the writes immediately before binding the
descriptors.

Fixes all of the validation messages regarding invalid command buffers
due to updated/deleted descriptor-sets.