Decouple data save paths from Emulator
This change allows the frontend to select the appdata and configuration
file locations when instantiating the emulator, allowing for more
flexibility in placing the files.

This also fixes configuration not working in MacOS app bundles,
as it will now look for config.toml in the user's "Application Support"
directory and default there.